refactor(api): put metrics & healthcheck in instance file
